+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 3.6.1
 * `FIX` wrong diagnostics for `pcall` and `xpcall`
+* `FIX` duplicate fields in table hover
 
 ## 3.6.0
 `2022-11-8`